before new bin pileup flattener
Former-commit-id: 4cfed80cb061041128bdbbd9b608d77bfecd95ce

+++ branches/new_data_structures/GenPlay/src/edu/yu/einstein/genplay/core/IO/extractor/ElandExtendedExtractor.java
@@ -167,7 +167,7 @@ public final class ElandExtendedExtractor extends TextFileExtractor implements S
 			return LINE_SKIPPED;
 		}
 
-		chromoNumber = getProjectChromosome().getIndex(chromosomeName);
+		chromoNumber = (short) getProjectChromosome().getIndex(chromosomeName);
 		// try to extract the position number
 		i+=4;  // we want to get rid of 'fa:'
 		j = 0;
